+++ docs/CHANGELOG.md
@@ -323,4 +323,4 @@ START OF THE PROJECT
 **v. 3.3.4 (xx.03.23)**
 - Allowed any item to be deserialized into object (it's for supporting map of multiple values)
 - Added support for serialization using ExpandoObject
-- //TODO generate schema for map and nullable record
+- Extended GenerateModel to support map of union type